Updated to use version Az Functions, v3. Also, created PowerShell scripts for the equivalent BASH scripts.

+++ src/deployment/create-file-list.sh
@@ -1,10 +1,11 @@
-#!/bin/bash
-
 SOURCE_FOLDER='../portal/'
 DESTINATION_FOLDER='$web/'
 TERRAFORM_FILE='web-file-list.tf'
 EMPTY_STRING=''
-rm $TERRAFORM_FILE
+
+if [[ -f $TERRAFORM_FILE ]]; then
+    rm $TERRAFORM_FILE
+fi
 
 exec 1>>$TERRAFORM_FILE
